Start on declaring the behaviors for various prototypes in descriptors.lua, to avoid if tree pain, and use it for inserter read modes
As the comments say, most things are pretty generic.  This avoids duplicating them and, shortly, will kill a whole lot of circuit-networks.lua
as well as letting us finish supporting all the circuit network things.  For thew time being, we limit the fallout to inserters, as we did
with the circular options list version.  Now, though, it should be possible to see what the end
state looks like, both for circuit networks specifically and additionally for lots of other things.